requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
textblob>=0.17.0
firebase-admin>=6.0.0
requests>=2.31.0

